feat: parse chapter titles from navigation documents
