+++ epub/package.go
@@ -1,0 +1,138 @@
+package epub
+
+import (
+	"archive/zip"
+	"encoding/xml"
+	"io"
+	"path"
+)
+
+// Package represents an epub .opf file.
+type Package struct {
+	Metadata
+	Manifest
+	Spine
+}
+
+// Metadata contains publishing information about the epub.
+type Metadata struct {
+	Title      string `xml:"metadata>title"`
+	Language   string `xml:"metadata>language"`
+	Identifier struct {
+		Scheme  string `xml:"scheme,attr"`
+		Content string `xml:",innerxml"`
+	} `xml:"metadata>identifier"`
+	Creator     string `xml:"metadata>creator"`
+	Contributor string `xml:"metadata>contributor"`
+	Publisher   string `xml:"metadata>publisher"`
+	Subject     string `xml:"metadata>subject"`
+	Description string `xml:"metadata>description"`
+	Dates       []struct {
+		Event string `xml:"event,attr"`
+		Date  string `xml:",innerxml"`
+	} `xml:"metadata>date"`
+	Type     string `xml:"metadata>type"`
+	Format   string `xml:"metadata>format"`
+	Source   string `xml:"metadata>source"`
+	Relation string `xml:"metadata>relation"`
+	Coverage string `xml:"metadata>coverage"`
+	Rights   string `xml:"metadata>rights"`
+}
+
+// Manifest lists every file that is part of the epub.
+type Manifest struct {
+	Items []Item `xml:"manifest>item"`
+}
+
+// Item represents a file stored in the epub.
+type Item struct {
+	ID        string `xml:"id,attr"`
+	HREF      string `xml:"href,attr"`
+	MediaType string `xml:"media-type,attr"`
+	Label     string
+	f         *zip.File
+}
+
+// Open returns a ReadCloser that provides access to the Items's contents.
+// Multiple items may be read concurrently.
+func (item *Item) Open() (r io.ReadCloser, err error) {
+	if item.f == nil {
+		return nil, ErrBadManifest
+	}
+
+	return item.f.Open()
+}
+
+// Close closes the epub file, rendering it unusable for I/O.
+func (rc *ReadCloser) Close() {
+	rc.f.Close()
+}
+
+// Spine defines the reading order of the epub documents.
+type Spine struct {
+	Itemrefs []Itemref `xml:"spine>itemref"`
+}
+
+// Itemref points to an Item.
+type Itemref struct {
+	IDREF string `xml:"idref,attr"`
+	*Item
+}
+
+// setPackages unmarshal's each of the epub's .opf files.
+func (r *Reader) setPackages() error {
+	for _, rf := range r.Container.Rootfiles {
+		if r.files[rf.FullPath] == nil {
+			return ErrBadRootfile
+		}
+
+		f, err := r.files[rf.FullPath].Open()
+		if err != nil {
+			return err
+		}
+		defer f.Close()
+
+		data, err := io.ReadAll(f)
+		if err != nil {
+			return err
+		}
+
+		err = xml.Unmarshal(data, &rf.Package)
+		if err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
+
+// setItems associates Itemrefs with their respective Item and Items with
+// their zip.File.
+func (r *Reader) setItems() error {
+	itemrefCount := 0
+	for _, rf := range r.Container.Rootfiles {
+		itemMap := make(map[string]*Item)
+		for i := range rf.Manifest.Items {
+			item := &rf.Manifest.Items[i]
+			itemMap[item.ID] = item
+
+			abs := path.Join(path.Dir(rf.FullPath), item.HREF)
+			item.f = r.files[abs]
+		}
+
+		for i := range rf.Spine.Itemrefs {
+			itemref := &rf.Spine.Itemrefs[i]
+			itemref.Item = itemMap[itemref.IDREF]
+			if itemref.Item == nil {
+				return ErrBadItemref
+			}
+		}
+		itemrefCount += len(rf.Spine.Itemrefs)
+	}
+
+	if itemrefCount < 1 {
+		return ErrNoItemref
+	}
+
+	return nil
+}
